Things went wrong so that was fun.
So a lot of things went wrong so I just deleted all the code for the lift movement. Also I started preparing for having two Lift motors, but we didn't have the converter cable. Also created a ServoCallibration class to do measurements for the grabbers and the lift. Autonomous testing did kind of work though so that was cool.